seperates concerns load screen
